% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.R
\name{rep_c}
\alias{rep_c}
\title{Repeat a vector}
\usage{
rep_c(x, times)
}
\arguments{
\item{x}{A single numeric value}

\item{times}{A single numeric value}
}
\description{
\code{rep_c()} repeats a vector
}
\examples{
rep_c(1, 3)

}
